+++ tests/Encoders.fs
@@ -690,7 +690,10 @@ let tests : Test =
                 equal expected actual
 
             testCase "Encode.Auto.generateEncoder throws for field using a non optional class" <| fun _ ->
-                let expected = "Cannot generate auto encoder for Tests.Types.BaseClass. Please pass an extra encoder."
+                let expected = """Cannot generate auto encoder for Tests.Types.BaseClass. Please pass an extra encoder.
+
+Documentation available at: https://thoth-org.github.io/Thoth.Json/documentation/auto/extra-coders.html#ready-to-use-extra-coders"""
+
                 let errorMsg =
                     try
                         let encoder = Encode.Auto.generateEncoder<RecordWithRequiredClass>(caseStrategy = CamelCase)